requirements.txt
# ==============================================================================
# Core ML / Training
# ==============================================================================
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
transformers>=4.44.0
datasets>=2.21.0
accelerate>=0.33.0
peft>=0.12.0
trl>=0.10.0
bitsandbytes>=0.43.0
safetensors>=0.4.0
tokenizers>=0.19.0
sentencepiece
protobuf

# ==============================================================================
# Distributed Training
# ==============================================================================
deepspeed>=0.14.0
# flash-attn>=2.6.0  # Install separately: pip install flash-attn --no-build-isolation

# ==============================================================================
# Data Curation
# ==============================================================================
aiohttp>=3.10.0
aiofiles>=24.1.0
beautifulsoup4>=4.12.0
selectolax>=0.3.0
xxhash>=3.4.0
langdetect>=1.0.9
arxiv>=2.1.0

# ==============================================================================
# Model Merging
# ==============================================================================
# mergekit  # Optional: install from git for CLI merging

# ==============================================================================
# Evaluation
# ==============================================================================
rouge-score>=0.1.2
bert-score>=0.3.13
scipy>=1.14.0
lm-eval>=0.4.0

# ==============================================================================
# Inference / Serving
# ==============================================================================
vllm>=0.5.5
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
autoawq>=0.2.0
# auto-gptq>=0.7.0  # Optional: for GPTQ quantization

# ==============================================================================
# Monitoring / Observability
# ==============================================================================
wandb>=0.17.0
pynvml>=11.5.0
psutil>=6.0.0
prometheus-client>=0.20.0

# ==============================================================================
# API Clients (for LLM judge)
# ==============================================================================
openai>=1.40.0
anthropic>=0.34.0
httpx>=0.27.0

# ==============================================================================
# Utilities
# ==============================================================================
pyyaml>=6.0.0
pydantic>=2.8.0
tqdm>=4.66.0
numpy>=1.26.0
rich>=13.7.0
click>=8.1.0

# ==============================================================================
# Testing
# ==============================================================================
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# ==============================================================================
# Load Testing
# ==============================================================================
locust>=2.29.0
